Server/requirement.txt
FastAPI
pydantic
uvicorn
python-dotenv
passlib
pyjwt
motor
pymongo
pydantic
python-multipart
pydantic-settings
toon